+++ RedPandaIDE/translations/RedPandaIDE_pt_BR.ts
@@ -2001,6 +2001,18 @@
         <source>kb</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>Convert HTML for：</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Input</source>
+        <translation type="unfinished">Entrada</translation>
+    </message>
+    <message>
+        <source>Expected Output</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>FileAssociationModel</name>